Split task to count SVs per genome
Problem:
One of the more significant bottlenecks in the outlier exclusion
workflow is the step to count the number of SVs per genome from the
joined raw calls VCF. Parsing a compressed VCF file is a fairly
expensive operation and converting the entire file into a tab-delimited
format, as was done previously, is unnecessary if only a few fields are
used in downstream processing. In addition, counting the number of SVs
per genome after extracting the necessary information from the VCF can
be done more quickly using specialized tools instead of the
standard Linux utilites.

Solution:
- Split up the task to count SVs per genome into two tasks: one to
  extract the necessary information from the joined raw calls VCF and
  another to do the counting. This is mostly to help with caching.
  Having to repeatedly parse a large VCF due to workflow failures will
  be painful.
- Use bcftools query instead of svtk vcf2bed to parse the VCF and
  produce a table of SVs in each sample. bcftools query is more
  efficient given that only the SV types, SV lengths, sample IDs and
  variant IDs are needed.
- Use duckdb to count the SVs per genome. It is faster than a
  sort | uniq -c invocation and the database can be used for other
  queries downstream.