Include user SV type in filter
Problem:
When the joined raw calls variant database is created, only DELs,
DUPs, INVs, and INSs are included. If the user-specified SV type is not
one of these, those SVs will not be included in the database and will
not be used to determine outlier samples.

Solution:
Check if the SV type given by the user is equal to one of the retained
types and if it is not, add it to the filter to keep in the database.