Fix outlier variant filter labeling
Using bcftools annotate to annotate variants by ID with a tab-separated
file requires the REF and ALT columns to also be in the annotation file.
Otherwise, variants will be matched by CHROM and POS which may cause
some variants to be incorrectly annotated if they share a position with
a variant in the annotation file.