Do not use task input variable in outputs
Problem:
Passing a File as a task input, modifying the file and then using the
input variable name in the task output will not cause the modified file
to be delocalized. Instead, the original input path will be used as the
output path and the work done in the task will be lost.

Solution:
Create a local filename in the task command section and use that
filename as the task output.