use outlier samples file in conditional
If a table of outlier samples is given as input, the
ConvertVcfOrBcfToTsv, MakeSvCountsDb, and DetermineOutlierSamples tasks
should not execute. The first tasks are conditioned on the outlier
samples file not being defined while the last task is conditioned on the
existeance of the output of the first task. It seems that Cromwell
treats the output of scattered task inside a conditional as an empty
array and will both treat the output as being defined and pass an empty
array as input to any task that uses it. So in this case, if
ConvertVcfOrBcfToTsv is not executed, MakeSvCountsDb will still be
executed with an empty array of TSVs was input and thus fail. A
workaround is to make all three tasks conditional on the outlier samples
file being defined.